+++ R/read_csv.R
@@ -41,6 +41,10 @@
 #' For [sampling][model-method-sample] the returned list also includes the
 #' following components:
 #'
+#' * `time`: Run time information for the individual chains. The returned object
+#' is the same as for the [$time()][fit-method-time] method except the total run
+#' time can't be inferred from the CSV files (the chains may have been run in
+#' parallel) and is therefore `NA`.
 #' * `inv_metric`: A list (one element per chain) of inverse mass matrices
 #' or their diagonals, depending on the type of metric used.
 #' * `step_size`: A list (one element per chain) of the step sizes used.
@@ -127,6 +131,8 @@ read_cmdstan_csv <- function(files,
   step_size <- list()
   col_types <- NULL
   col_select <- NULL
+  metadata <- NULL
+  time <- data.frame()
   not_matching <- c()
   for (output_file in files) {
     if (is.null(metadata)) {
@@ -137,7 +143,9 @@ read_cmdstan_csv <- function(files,
       if (!is.null(metadata$step_size_adaptation)) {
         step_size[[as.character(metadata$id)]] <- metadata$step_size_adaptation
       }
-      id <- metadata$id
+      if (!is.null(metadata$time)) {
+        time <- rbind(time, metadata$time)
+      }
     } else {
       csv_file_info <- read_csv_metadata(output_file)
       check <- check_csv_metadata_matches(metadata, csv_file_info)
@@ -158,7 +166,9 @@ read_cmdstan_csv <- function(files,
       if (!is.null(csv_file_info$step_size_adaptation)) {
         step_size[[as.character(csv_file_info$id)]] <- csv_file_info$step_size_adaptation
       }
-      id <- csv_file_info$id
+      if (!is.null(csv_file_info$time)) {
+        time <- rbind(time, csv_file_info$time)
+      }
     }
     if (is.null(col_select)) {
       if (is.null(variables)) { # variables = NULL returns all
@@ -328,6 +338,7 @@ read_cmdstan_csv <- function(files,
     }
     list(
       metadata = metadata,
+      time = list(total = NA_integer_, chains = time),
       inv_metric = inv_metric,
       step_size = step_size,
       warmup_draws = warmup_draws,
@@ -510,6 +521,9 @@ read_csv_metadata <- function(csv_file) {
   inv_metric_rows <- -1
   parsing_done <- FALSE
   dense_inv_metric <- FALSE
+  warmup_time <- 0
+  sampling_time <-0
+  total_time <- 0
   if (os_is_windows()) {
     grep_path <- repair_path(Sys.which("grep.exe"))
     fread_cmd <- paste0(grep_path, " '^[#a-zA-Z]' --color=never ", csv_file)
@@ -584,6 +598,16 @@ read_csv_metadata <- function(csv_file) {
               csv_file_info[[key_val[1]]] <- key_val[2]
             }
           }
+        } else if (grepl("(Warm-up)", tmp, fixed = TRUE)) {
+          tmp <- gsub("Elapsed Time:", "", tmp, fixed = TRUE)
+          tmp <- gsub("seconds (Warm-up)", "", tmp, fixed = TRUE)
+          warmup_time <- as.numeric(tmp)
+        } else if (grepl("(Sampling)", tmp, fixed = TRUE)) {
+          tmp <- gsub("seconds (Sampling)", "", tmp, fixed = TRUE)
+          sampling_time <- as.numeric(tmp)
+        } else if (grepl("(Total)", tmp, fixed = TRUE)) {
+          tmp <- gsub("seconds (Total)", "", tmp, fixed = TRUE)
+          total_time <- as.numeric(tmp)
         }
       }
     }
@@ -610,6 +634,14 @@ read_csv_metadata <- function(csv_file) {
     csv_file_info$threads <- csv_file_info$num_threads
   } else {
     csv_file_info$threads_per_chain <- csv_file_info$num_threads
+  }
+  if (csv_file_info$method == "sample") {
+    csv_file_info$time <- data.frame(
+      chain_id = csv_file_info$id,
+      warmup = warmup_time,
+      sampling = sampling_time,
+      total = total_time
+    )
   }
   csv_file_info$model <- NULL
   csv_file_info$engaged <- NULL